chore: Add relationship between Listing and User models

+++ app/Http/Controllers/ListingController.php
@@ -44,6 +44,8 @@ class ListingController extends Controller {
             $formFields['logo'] = $request->file('logo')->store('logos', 'public');
         }
 
+        $formFields['user_id'] = auth()->id();
+
         Listing::create($formFields);
 
         return redirect('/')->with('message', 'Listing created successfully!');
@@ -56,8 +58,13 @@ class ListingController extends Controller {
         ]);
     }
 
-    // Store listing data
+    // Update listing data
     public function update(Request $request, Listing $listing) {
+        // Make sure logged in user is the owner of the listing
+        if ($listing->user_id !== auth()->id()) {
+            abort(403, 'Unauthorized action.');
+        }
+
         $formFields = $request->validate([
             'title' => 'required',
             'company' => 'required',
@@ -79,8 +86,20 @@ class ListingController extends Controller {
 
     // Delete listing
     public function destroy(Listing $listing) {
+        // Make sure logged in user is the owner of the listing
+        if ($listing->user_id !== auth()->id()) {
+            abort(403, 'Unauthorized action.');
+        }
+
         $listing->delete();
 
         return redirect('/')->with('message', 'Listing deleted successfully!');
+    }
+
+    // Manage listings
+    public function manage() {
+        return view('listings.manage', [
+            'listings' => auth()->user()->listings()->get(),
+        ]);
     }
 }
